sprint animation + tails fly ability
When velocity.x is over or equal to 400, the animation changes from running to sprinting. However, this would only work in 1 direction, the direction which the velocity is positive (i think right) since when moving in the other direction, the velocity is multiplied by the direction (-1) so it turns negative to apply veloctiy in the CORRECT direction. This creates a problem since the math wont work since we will ALWAYS be below 400 velocity when moving in the negative velocity direction. Issue was solved by using abs() function, which gets the absolute value, without the negative.

Did same for tails, but there isn't a sprint animation for him (yet, there probably is on a sprite sheet) so its just if below 200 velocity or above 200 play run animation. In addition, tails can toggle flying with F key.

Long ass description.

I think this is the last commit of the branch, will merge back into main after.

Co-Authored-By: Niscoro <69369889+Niscoro@users.noreply.github.com>