LimboConsole, 2 Console Commands
Added LimboConsole plugin/addon and added to commands for demonstration and debugging purposes, "health" and "debugging". Health command takes 2 parameters, first parameter is for the character's health you want to modify and the second parameter is for the value you want to modify it by.

Ideally, we'd have a command called "stat" or "setstat" or "set" or something which could go like this:

set stat hp sonic 50
which would set sonic's HP stat to 50:
or we could just keep it simple with commands like

hp sonic 50
or
sethp sonic 50
I think that's better.

"Debugging" command just takes one parameter, a boolean. If true, debugMode is activated and vice versa.
Could also add a command for a debugMenu which could do the same thing as the commands but with a visual interface.