Command Autocompletion, Feedback, More Commands
Command Autocompletion long story short, yk how arrays start at 0? Thats the same thing with parameters in these commands, the first parameter in command which is like the command name so if we were in minecraft:
/give MasterKuby gold_block 50

The "give" part would be the 0th, then the "MasterKuby" part would be the 1st and so on... kinda shitty explanation but you'll understand.

So anyways, the parameters in script of the autocomplete thing are:
First Param: Command Name
Second Param: Command Parameter Slot/ID
Third Param: Command Autocomplete Options

And as for command feedback, LimboConsole.info basically just prints a string to the LimboConsole console. You can put some variables in there as long as they're stringified (bro did not say "stringified").

The additional command is the "debugmode" command, does the EXACT same thing as the "debugging" command but somebody may want it like that.

Also changed the command function names in script to like include "Command" at the end just for organisations sake. Is that a good idea or not?