Composition | HP Component & Hitbox Component
Added a HP Component and a Hitbox Component which are their own classes and can be added as nodes using the "Add/Create a node" thing. The components are quite literally plug and play, the only thing I can think of that will need tinkering in the parent's script is for the logic when the parent dies. How the Hitbox Component will work is that when a Damage Component collides with the Hitbox Component, the hit() function in the Hitbox Component will be called by the Damage Component by the Damage Component getting the node it collided with. The Damage Component will have an externally configurable damage variable (configurable in the editor). The hit function has one parameter (the damage value) however in the future it may have more parameters such as a parameter for an effect (fire, poison etc). The hit() function calls the takeDamage() function in the HP Component, carrying over the damage value. The takeDamage function then subtracts the value of the damage variable by the HP. If the parent of the HP Component is a character, then that character's global HP is set to the HP Component's HP. The HP Component has a configurable HP Value (configurable in editor). The HP Component will also update the local HP with the global HP when a change in the global value is detected. All of this is basically plug and play because there is almost no setup required in the parent. Hitbox Component depends on the HP Component. Hitbox Component also obviously needs a collision shape since its just an extension of an Area2D. You must assign a HP Component to the Hitbox Component.

Damage Component will prolly be added next commit, you'd just attach it to your projectile and call it a day. I think the Damage Component will call a function in the parent (the projectile) when it hits something, so the bullet can do various things (queue_free, object pool, hide etc)

In addition, "health" has been changed to "HP" project-wide.

Bye!